+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Provide bulk copybook download via Zowe Explorer
 - Support program and assembler type symbol attributes
 - Support SYSATTRA and SYSATTRP functions
+- Show program and assembler type attributes in hover text
 
 #### Fixed
 - Logs produced by language server in WebAssembly mode are not visible